integrate generator-ts.mts
